requirements.txt
django
djangorestframework
drf-yasg
numpy
pandas
sklearn
gunicorn
pillow
polls
markdown
django-filter
